requirements.txt
Keras>=2.3.1
tensorflow-gpu>=1.13.1
numpy>=1.15.0
opencv-python>=4.2.0
scikit-learn>=0.21.2
scipy>=1.4.1
Pillow
torch>=1.3.0
torchvision>=0.4.1
natsort>=8.1.0